+++ workspaces/theme/plugins/theme/src/darkTheme.test.ts
@@ -14,15 +14,11 @@
  * limitations under the License.
  */
 
-import { customDarkTheme } from './darkTheme';
+import { customDarkTheme, darkThemeOverrides } from './darkTheme';
 
 describe('customDarkTheme', () => {
   it('should return the correct defaults for dark mode', () => {
     expect(customDarkTheme()).toEqual({
-      background: {
-        default: '#333333',
-        paper: '#1b1d21',
-      },
       banner: {
         closeButtonColor: '#FFFFFF',
         error: '#E22134',
@@ -51,27 +47,9 @@ describe('customDarkTheme', () => {
       link: '#9CC9FF',
       linkHover: '#82BAFD',
       mode: 'dark',
-      navigation: {
-        background: '#0f1214',
-        indicator: '#0066CC',
-        color: '#ffffff',
-        selectedColor: '#ffffff',
-        navItem: {
-          hoverBackground: '#3c3f42',
-        },
-        submenu: {
-          background: '#0f1214',
-        },
-      },
       pinSidebarButton: {
         background: '#BDBDBD',
         icon: '#404040',
-      },
-      primary: {
-        main: '#1FA7F8',
-      },
-      secondary: {
-        main: '#1FA7F8',
       },
       status: {
         aborted: '#9E9E9E',
@@ -91,55 +69,7 @@ describe('customDarkTheme', () => {
       warningBackground: '#F59B23',
       warningText: '#000000',
 
-      rhdh: {
-        general: {
-          disabled: '#AAABAC',
-          disabledBackground: '#444548',
-
-          paperBackgroundImage: 'none',
-          paperBorderColor: '#A3A3A3',
-
-          popoverBoxShadow:
-            '0 0.25rem 0.5rem 0rem rgba(3, 3, 3, 0.48), 0 0 0.25rem 0 rgba(3, 3, 3, 0.24)',
-
-          cardBackgroundColor: '#1b1d21',
-          cardBorderColor: '#A3A3A3',
-
-          headerBottomBorderColor: '#A3A3A3',
-          mainSectionBackgroundColor: '#0f1214',
-          formControlBackgroundColor: '#36373A',
-
-          sidebarBackgroundColor: '#1b1d21',
-          sidebarItemSelectedBackgroundColor: '#4F5255',
-
-          tableTitleColor: '#E0E0E0',
-          tableSubtitleColor: '#E0E0E0',
-          tableColumnTitleColor: '#E0E0E0',
-          tableRowHover: '#0f1214',
-          tableBorderColor: '#515151',
-          tableBackgroundColor: '#1b1d21',
-          tabsDisabledBackgroundColor: '#444548',
-          tabsBottomBorderColor: '#444548',
-
-          contrastText: '#FFF',
-
-          appBarBackgroundColor: '#1b1d21',
-          appBarBackgroundImage: 'none',
-        },
-        primary: {
-          main: '#1FA7F8',
-          focusVisibleBorder: '#ADD6FF',
-        },
-        secondary: {
-          main: '#1FA7F8',
-          focusVisibleBorder: '#ADD6FF',
-        },
-        cards: {
-          headerTextColor: '#FFF',
-          headerBackgroundColor: '#0f1214',
-          headerBackgroundImage: 'none',
-        },
-      },
+      ...darkThemeOverrides,
     });
   });
 });