+++ workspaces/theme/plugins/theme/src/lightTheme.test.ts
@@ -14,15 +14,11 @@
  * limitations under the License.
  */
 
-import { customLightTheme } from './lightTheme';
+import { customLightTheme, lightThemeOverrides } from './lightTheme';
 
 describe('customLightTheme', () => {
   it('should return the correct defaults for light mode', () => {
     expect(customLightTheme()).toEqual({
-      background: {
-        default: '#F8F8F8',
-        paper: '#FFFFFF',
-      },
       banner: {
         closeButtonColor: '#FFFFFF',
         error: '#E22134',
@@ -51,27 +47,9 @@ describe('customLightTheme', () => {
       link: '#0A6EBE',
       linkHover: '#2196F3',
       mode: 'light',
-      navigation: {
-        background: '#222427',
-        indicator: '#0066CC',
-        color: '#ffffff',
-        selectedColor: '#ffffff',
-        navItem: {
-          hoverBackground: '#3c3f42',
-        },
-        submenu: {
-          background: '#222427',
-        },
-      },
       pinSidebarButton: {
         background: '#BDBDBD',
         icon: '#181818',
-      },
-      primary: {
-        main: '#0066CC',
-      },
-      secondary: {
-        main: '#0066CC',
       },
       status: {
         aborted: '#757575',
@@ -90,60 +68,7 @@ describe('customLightTheme', () => {
       type: 'light',
       warningBackground: '#F59B23',
       warningText: '#000000',
-
-      text: {
-        primary: '#151515',
-        secondary: '#757575',
-      },
-      rhdh: {
-        general: {
-          disabled: '#6A6E73',
-          disabledBackground: '#D2D2D2',
-
-          paperBackgroundImage: 'none',
-          paperBorderColor: '#C7C7C7',
-
-          popoverBoxShadow:
-            '0 0.25rem 0.5rem 0rem rgba(3, 3, 3, 0.12), 0 0 0.25rem 0 rgba(3, 3, 3, 0.06)',
-
-          cardBackgroundColor: '#FFF',
-          cardBorderColor: '#C7C7C7',
-
-          headerBottomBorderColor: '#C7C7C7',
-          mainSectionBackgroundColor: '#FFF',
-          formControlBackgroundColor: '#FFF',
-
-          sidebarBackgroundColor: '#212427',
-          sidebarItemSelectedBackgroundColor: '#4F5255',
-
-          tableTitleColor: '#181818',
-          tableSubtitleColor: '#616161',
-          tableColumnTitleColor: '#151515',
-          tableRowHover: '#F5F5F5',
-          tableBorderColor: '#E0E0E0',
-          tableBackgroundColor: '#FFF',
-          tabsDisabledBackgroundColor: '#f5f5f5',
-          tabsBottomBorderColor: '#D2D2D2',
-
-          contrastText: '#FFF',
-
-          appBarBackgroundColor: '#212427',
-          appBarBackgroundImage: 'none',
-        },
-        primary: {
-          main: '#0066CC',
-          focusVisibleBorder: '#0066CC',
-        },
-        secondary: {
-          main: '#0066CC',
-          focusVisibleBorder: '#0066CC',
-        },
-        cards: {
-          headerTextColor: '#151515',
-          headerBackgroundColor: '#FFF',
-          headerBackgroundImage: 'none',
-        },
-      },
+      ...lightThemeOverrides,
     });
   });
 });